cement_ai_platform/cement-plant/requirements.txt
fastapi
pydantic
uvicorn
joblib
pandas
numpy
requests
google-adk
google-genai
python-dotenv
scikit-learn
